Update helpers.py numpy product call since deprecated
NumPy deprecated numpy.Product in v1.25 since it just called numpy.prod underneath. New numpy version give a deprecation warning for calling it and it will be removed in numpy v2.0. Updating the call to be correct. Since Product just called prod there should be no change in operation of the code.